Setup for X - Words
X will be a simple game where there's a five letter word, one of which is missing, and you have to toggle which letter given you think is correct before the timer locks your answer in.  Right now, I mostly focused on preparing the UI and the list of words that will eventually need to be populated with a LOT of words.  At least I know what I'm working on tomorrow.